DOM and Use of JS file within a HTML document and Providing  Source of JS File
Learned of the importance of the document object model as it allows to make sense of the hierarchy of the HTML document and as a result be able to manipulate the attributes of the objects involved. Lastly, I was able to learn of how to either write the js script within the HTML document or include the name of the js source file.